🎨 refactor: prompts to utils 프롬프트 리펙토링
