+++ app/utils/prompt/agent_prompt.py
@@ -1,7 +1,7 @@
 """프롬프트 템플릿 정의."""
 from langchain_core.prompts import ChatPromptTemplate, MessagesPlaceholder
 
-SYSTEM_PROMPT = """당신은 유용한 AI 어시스턴트입니다. 사용자의 질문에 답하기 위해 필요할 때마다 도구를 적극적으로 사용하세요.
+AGENT_PROMPT = """당신은 유용한 AI 어시스턴트입니다. 사용자의 질문에 답하기 위해 필요할 때마다 도구를 적극적으로 사용하세요.
 
 특히 다음과 같은 경우에는 반드시 도구를 사용해야 합니다:
 1. 실시간 정보가 필요한 경우 (날씨, 뉴스 등)
@@ -44,6 +44,6 @@ SYSTEM_PROMPT = """당신은 유용한 AI 어시스턴트입니다. 사용자의
 
 # 변경된 부분: messages와 scratchpad를 하나의 MessagesPlaceholder로 통합
 prompt = ChatPromptTemplate.from_messages([
-    ("system", SYSTEM_PROMPT),
+    ("system", AGENT_PROMPT),
     MessagesPlaceholder(variable_name="messages"),  # 이전 대화 기록과 현재 입력을 포함
 ])
