+++ app/utils/prompt/insight_prompts.py
@@ -1,0 +1,27 @@
+# 인사이트 생성용 LLM 프롬프트 템플릿 모음
+
+INSIGHT_ANALYSIS_PROMPT = """
+너는 사용자의 키워드에 기반하여 수집된 최신 웹 정보에서 인사이트를 도출해야 해.
+단순 요약이 아니라, 키워드 간 연관성과 새로운 기회, 흐름, 의미 있는 관찰을 종합해줘.
+키워드:
+{keywords}
+웹 문서 요약:
+{context}
+[요구사항]
+- 키워드 간 연결 구조를 파악해
+- 트렌드나 기회가 보이면 제안해
+- 사용자에게 통찰이나 질문을 던져
+결과:
+"""
+
+INSIGHT_COMPARE_PROMPT = """
+[최근 뉴스 요약]
+{recent_summary}
+
+[과거 뉴스 요약]
+{past_summary}
+
+[요구사항]
+- 두 요약의 차이점, 연결점, 시사점, 트렌드 변화를 분석해줘
+- 독자에게 새로운 시각, 의문, 또는 비즈니스 아이디어를 제시해줘
+"""
